+++ include/SW_Weather.h
@@ -190,8 +190,6 @@ void SW_WTH_setWeatherValues(
     double ***tempWeather,
     double elevation,
     TimeInt doyOffset,
-    size_t *domSuid,
-    Bool sDom,
     SW_WEATHER_HIST *yearlyWeather,
     LOG_INFO *LogInfo
 );
@@ -233,8 +231,6 @@ void finalizeAllWeather(
     SW_WEATHER_HIST *allHist,
     TimeInt cum_monthdays[],
     TimeInt days_in_month[],
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 );
 
@@ -261,17 +257,11 @@ void generateMissingWeather(
     unsigned int n_years,
     unsigned int method,
     unsigned int optLOCF_nMax,
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 );
 
 void checkAllWeather(
-    SW_WEATHER_INPUTS *weather,
-    SW_WEATHER_HIST *weathHist,
-    size_t ncSuid[],
-    Bool sDom,
-    LOG_INFO *LogInfo
+    SW_WEATHER_INPUTS *weather, SW_WEATHER_HIST *weathHist, LOG_INFO *LogInfo
 );
 
 void SW_WTH_allocateAllWeather(
@@ -292,8 +282,6 @@ void SW_WTH_finalize_all_weather(
     SW_WEATHER_HIST *allHist,
     TimeInt cum_monthdays[],
     TimeInt days_in_month[],
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 );