+++ src/SW_Weather.c
@@ -102,25 +102,13 @@
 
 
 static void helperPercentValue(
-    double *val,
-    Bool fixPERCENT,
-    const char *name,
-    TimeInt year,
-    TimeInt doy,
-    size_t *domSuid,
-    Bool sDom,
-    LOG_INFO *LogInfo
+    double *val, Bool fixPERCENT, const char *name, LOG_INFO *LogInfo
 ) {
     if (!missing(*val) && *val > 100.) {
-        LogErrorSuid(
+        LogError(
             LogInfo,
             fixPERCENT ? LOGWARN : LOGERROR,
-            domSuid,
-            sDom,
-            fixPERCENT ? "%d-%03d: Reset %s (%f) to 100%%." :
-                         "%d-%03d: %s (%f) > 100%%.",
-            year,
-            doy,
+            fixPERCENT ? "Reset %s (%f) to 100%%." : "%s (%f) > 100%%.",
             name,
             *val
         );
@@ -763,9 +751,6 @@ transfer/calculate to SW_WEATHER_HIST for the simulation
 @param[in] doyOffset A value to offset yearly days if the input array
     holds more that one site's worth of values; should be set to 0 if
     SWMPI is not enabled
-@param[in] domSuid Suid indices of the program-domain for the current
-    input instance being set (MPI only)
-@param[in] sDom Specifies the program's domain is site-oriented (MPI/NC only)
 @param[out] yearlyWeather Destination for temporary/calculated values
 for all years within the simulation
 @param[out] LogInfo Holds information on warnings and errors
@@ -778,8 +763,6 @@ void SW_WTH_setWeatherValues(
     double ***tempWeather,
     double elevation,
     TimeInt doyOffset,
-    size_t *domSuid,
-    Bool sDom,
     SW_WEATHER_HIST *yearlyWeather,
     LOG_INFO *LogInfo
 ) {
@@ -803,13 +786,10 @@ void SW_WTH_setWeatherValues(
 
     if (useHumidityDaily && !useMaxMinRelHumid && !inputFlags[REL_HUMID] &&
         inputFlags[SPEC_HUMID] && missing(elevation)) {
-        LogErrorSuid(
+        LogError(
             LogInfo,
             LOGERROR,
-            domSuid,
-            sDom,
-            "Elevation is missing but required to "
-            "calculate relative humidity "
+            "Elevation is missing but required to calculate relative humidity "
             "from specific humidity."
         );
 
@@ -820,6 +800,13 @@ void SW_WTH_setWeatherValues(
 
         for (doy = 0; doy < MAX_DAYS; doy++) {
             tempDoy = doy + doyOffset;
+
+            formatLogDate(
+                LogInfo->logDate,
+                sizeof LogInfo->logDate,
+                startYear + yearIndex,
+                doy + 1
+            );
 
             // Temperature [C]
             yearlyWeather[yearIndex].temp_max[doy] =
@@ -841,15 +828,10 @@ void SW_WTH_setWeatherValues(
                     yearlyWeather[yearIndex].temp_min[doy] >
                         yearlyWeather[yearIndex].temp_max[doy]) {
 
-                    LogErrorSuid(
+                    LogError(
                         LogInfo,
                         LOGWARN,
-                        domSuid,
-                        sDom,
-                        "%d-%03d: Swapped min/max air "
-                        "temperature %f/%f.",
-                        startYear + yearIndex,
-                        doy,
+                        "Swapped min/max air temperature %f/%f.",
                         yearlyWeather[yearIndex].temp_min[doy],
                         yearlyWeather[yearIndex].temp_max[doy]
                     );
@@ -878,10 +860,6 @@ void SW_WTH_setWeatherValues(
                     &yearlyWeather[yearIndex].cloudcov_daily[doy],
                     fixWeatherData[idFixPERCENT],
                     "cloud cover",
-                    startYear + yearIndex,
-                    doy,
-                    domSuid,
-                    sDom,
                     LogInfo
                 );
             }
@@ -925,16 +903,10 @@ void SW_WTH_setWeatherValues(
                                 tempWeather[yearIndex][REL_HUMID_MAX]
                                            [tempDoy]) {
 
-                            LogErrorSuid(
+                            LogError(
                                 LogInfo,
                                 LOGWARN,
-                                domSuid,
-                                sDom,
-                                "%d-%03d: "
-                                "Swapped min/max relative humidity "
-                                "%f/%f.",
-                                startYear + yearIndex,
-                                doy,
+                                "Swapped min/max relative humidity %f/%f.",
                                 tempWeather[yearIndex][REL_HUMID_MIN][tempDoy],
                                 tempWeather[yearIndex][REL_HUMID_MAX][tempDoy]
                             );
@@ -953,20 +925,12 @@ void SW_WTH_setWeatherValues(
                             &tempWeather[yearIndex][REL_HUMID_MAX][tempDoy],
                             fixWeatherData[idFixPERCENT],
                             "max relative humidity",
-                            startYear + yearIndex,
-                            doy,
-                            domSuid,
-                            sDom,
                             LogInfo
                         );
                         helperPercentValue(
                             &tempWeather[yearIndex][REL_HUMID_MIN][tempDoy],
                             fixWeatherData[idFixPERCENT],
                             "min relative humidity",
-                            startYear + yearIndex,
-                            doy,
-                            domSuid,
-                            sDom,
                             LogInfo
                         );
 
@@ -1021,10 +985,6 @@ void SW_WTH_setWeatherValues(
                     &yearlyWeather[yearIndex].r_humidity_daily[doy],
                     fixWeatherData[idFixPERCENT],
                     "relative humidity",
-                    startYear + yearIndex,
-                    doy,
-                    domSuid,
-                    sDom,
                     LogInfo
                 );
 
@@ -1121,10 +1081,6 @@ void SW_WTH_setWeatherValues(
                             &yearlyWeather[yearIndex].r_humidity_daily[doy],
                             fixWeatherData[idFixPERCENT],
                             "relative humidity",
-                            startYear + yearIndex,
-                            doy,
-                            domSuid,
-                            sDom,
                             LogInfo
                         );
                     }
@@ -1138,6 +1094,8 @@ void SW_WTH_setWeatherValues(
             }
         }
     }
+
+    LogInfo->logDate[0] = '\0';
 }
 
 /**
@@ -1344,8 +1302,6 @@ void readAllWeather(
             tempWeatherHist,
             elevation,
             0,
-            NULL,
-            swFALSE, // Does not matter
             allHist,
             LogInfo
         );
@@ -1366,9 +1322,6 @@ values related to temperature and weather generator
 @param[in] cum_monthdays Monthly cumulative number of days for "current"
 year
 @param[in] days_in_month Number of days per month for "current" year
-@param[in] ncSuid Unique indentifier of the current suid being simulated to
-    insert into a produced message (MPI or NC mode only)
-@param[in] sDom Specifies the program's domain is site-oriented (MPI/NC only)
 @param[out] LogInfo Holds information on warnings and errors
 
 Finalize weather values after they have been read in via
@@ -1381,11 +1334,8 @@ void finalizeAllWeather(
     SW_WEATHER_HIST *allHist,
     TimeInt cum_monthdays[],
     TimeInt days_in_month[],
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 ) {
-
     unsigned int day;
     unsigned int yearIndex;
 
@@ -1398,8 +1348,6 @@ void finalizeAllWeather(
             w->n_years,
             w->generateWeatherMethod,
             3, // optLOCF_nMax (TODO: make this user input)
-            ncSuid,
-            sDom,
             LogInfo
         );
         if (LogInfo->stopRun) {
@@ -1446,7 +1394,7 @@ void finalizeAllWeather(
 
     // Make sure all input, scaled, generated, and calculated daily weather
     // values are within reason
-    checkAllWeather(w, allHist, ncSuid, sDom, LogInfo);
+    checkAllWeather(w, allHist, LogInfo);
 }
 
 void SW_WTH_finalize_all_weather(
@@ -1455,8 +1403,6 @@ void SW_WTH_finalize_all_weather(
     SW_WEATHER_HIST *allHist,
     TimeInt cum_monthdays[],
     TimeInt days_in_month[],
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 ) {
 
@@ -1466,8 +1412,6 @@ void SW_WTH_finalize_all_weather(
         allHist,
         cum_monthdays,
         days_in_month,
-        ncSuid,
-        sDom,
         LogInfo
     );
 }
@@ -1675,9 +1619,6 @@ values related to temperature and weather generator
     missing values (see details).
 @param[in] optLOCF_nMax Maximum number of missing days per year (e.g., 5)
     before imputation by `LOCF` throws an error.
-@param[in] ncSuid Unique indentifier of the current suid being simulated to
-    insert into a produced message (MPI or NC mode only)
-@param[in] sDom Specifies the program's domain is site-oriented (MPI/NC only)
 @param[out] LogInfo Holds information on warnings and errors
 */
 void generateMissingWeather(
@@ -1687,8 +1628,6 @@ void generateMissingWeather(
     unsigned int n_years,
     unsigned int method,
     unsigned int optLOCF_nMax,
-    size_t ncSuid[],
-    Bool sDom,
     LOG_INFO *LogInfo
 ) {
 
@@ -1719,12 +1658,10 @@ void generateMissingWeather(
 
     // Error out if method not implemented
     if (method != wgLOCF && method != wgMKV) {
-        LogErrorSuid(
+        LogError(
             LogInfo,
             LOGERROR,
-            ncSuid,
-            sDom,
-            "generateMissingWeather(): method = %u is not implemented.\n",
+            "generateMissingWeather(): method = %u is not implemented.",
             method
         );
         return; // Exit function prematurely due to error
@@ -1738,8 +1675,12 @@ void generateMissingWeather(
         nFilledLOCF = 0;
 
         for (day = 0; day < numDaysYear; day++) {
-            /* Determine variables with missing values */
 
+            formatLogDate(
+                LogInfo->logDate, sizeof LogInfo->logDate, year, day + 1
+            );
+
+            /* Determine variables with missing values */
             missing_Tmax = (Bool) missing(allHist[yearIndex].temp_max[day]);
             missing_Tmin = (Bool) missing(allHist[yearIndex].temp_min[day]);
             missing_PPT = (Bool) missing(allHist[yearIndex].ppt[day]);
@@ -1772,8 +1713,6 @@ void generateMissingWeather(
                         SW_MarkovIn,
                         day,
                         year,
-                        ncSuid,
-                        sDom,
                         &allHist[yearIndex].temp_max[day],
                         &allHist[yearIndex].temp_min[day],
                         &allHist[yearIndex].ppt[day],
@@ -1839,14 +1778,12 @@ void generateMissingWeather(
                     }
 
                     if (nFilledLOCF > optLOCF_nMax) {
-                        LogErrorSuid(
+                        LogError(
                             LogInfo,
                             LOGERROR,
-                            ncSuid,
-                            sDom,
-                            "generateMissingWeather(): more than %u days "
-                            "missing in year %u "
-                            "and weather generator turned off.\n",
+                            "generateMissingWeather(): "
+                            "more than %u days missing in year %u "
+                            "and weather generator turned off.",
                             optLOCF_nMax,
                             year
                         );
@@ -1875,6 +1812,8 @@ void generateMissingWeather(
             }
         }
     }
+
+    LogInfo->logDate[0] = '\0';
 }
 
 /**
@@ -1886,18 +1825,10 @@ program crash.
 @param[in] weather Struct of type SW_WEATHER_INPUTS holding all relevant
 information pretaining to weather input data
 @param[in] weathHist Array containing all historical data of a site
-@param[in] ncSuid Unique indentifier of the current suid being simulated to
-    insert into a produced message (MPI or NC mode only)
-@param[in] sDom Specifies the program's domain is site-oriented (MPI or NC
-mode only)
 @param[out] LogInfo Holds information on warnings and errors
 */
 void checkAllWeather(
-    SW_WEATHER_INPUTS *weather,
-    SW_WEATHER_HIST *weathHist,
-    size_t ncSuid[],
-    Bool sDom,
-    LOG_INFO *LogInfo
+    SW_WEATHER_INPUTS *weather, SW_WEATHER_HIST *weathHist, LOG_INFO *LogInfo
 ) {
 
     // Initialize any variables
@@ -1914,6 +1845,12 @@ void checkAllWeather(
 
         // Loop through `allHist` days
         for (doy = 0; doy < numDaysInYear; doy++) {
+            formatLogDate(
+                LogInfo->logDate,
+                sizeof LogInfo->logDate,
+                year + weather->startYear,
+                doy + 1
+            );
 
             dailyMaxTemp = weathHist[year].temp_max[doy];
             dailyMinTemp = weathHist[year].temp_min[doy];
@@ -1923,19 +1860,12 @@ void checkAllWeather(
                 // Check if minimum temp greater than maximum temp
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Daily input value for minimum temperature"
-                    " is greater than daily input value for maximum "
-                    "temperature (minimum = %f, maximum = %f)"
-                    " on day %d of year %d.",
+                    "Daily minimum > maximum air temperature (%f > %f).",
                     dailyMinTemp,
-                    dailyMaxTemp,
-                    doy + 1,
-                    year + weather->startYear
+                    dailyMaxTemp
                 );
                 // Will exit function prematurely due to error
 
@@ -1946,17 +1876,13 @@ void checkAllWeather(
                 // dew point temp is not [-100, 100]
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Daily minimum and/or maximum temperature on "
-                    "day %d of year %d do not fit in the range of [-100, "
-                    "100] "
-                    "C.",
-                    doy,
-                    year + weather->startYear
+                    "Daily minimum and/or maximum air temperature "
+                    "outside [-100, 100] C (Tmin = %f, Tmax = %f)",
+                    dailyMinTemp,
+                    dailyMaxTemp
                 );
                 // Will exit function prematurely due to error
 
@@ -1965,17 +1891,11 @@ void checkAllWeather(
                 // Otherwise, check if precipitation is less than 0cm
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Invalid daily precipitation value: %f cm (< 0) on day "
-                    "%d "
-                    "of year %d.",
-                    weathHist[year].ppt[doy],
-                    doy + 1,
-                    year + weather->startYear
+                    "Invalid daily precipitation amount: %f cm (< 0).",
+                    weathHist[year].ppt[doy]
                 );
                 // Will exit function prematurely due to error
 
@@ -1987,28 +1907,20 @@ void checkAllWeather(
 
                 if ((weathHist[year].r_humidity_daily[doy] >= 0.) &&
                     (weathHist[year].r_humidity_daily[doy] <= 1.)) {
-                    LogErrorSuid(
+                    LogError(
                         LogInfo,
                         LOGWARN,
-                        ncSuid,
-                        sDom,
-                        "Daily/calculated relative humidity value (%f) is "
+                        "Daily/calculated relative humidity (%f) "
                         "within [0, 1] indicating a possibly incorrect "
-                        "unit "
-                        "(expectation: value within [0, 100] %%).",
+                        "unit (expectation: value within [0, 100] %%).",
                         weathHist[year].r_humidity_daily[doy]
                     );
                 } else {
-                    LogErrorSuid(
+                    LogError(
                         LogInfo,
                         LOGERROR,
-                        ncSuid,
-                        sDom,
-                        "Invalid daily/calculated relative humidity value "
-                        "did"
-                        " not fall in the range [0, 100] (relative "
-                        "humidity = "
-                        "%f). ",
+                        "Daily/calculated relative humidity outside "
+                        "[0, 100] (relative humidity = %f).",
                         weathHist[year].r_humidity_daily[doy]
                     );
                     return; // Exit function prematurely due to error
@@ -2021,13 +1933,11 @@ void checkAllWeather(
                 // if the value is less than 0% or greater than 100%
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Invalid daily/calculated cloud cover value did"
-                    " not fall in the range [0, 100] (cloud cover = %f). ",
+                    "Daily/calculated cloud cover outside [0, 100] "
+                    "(cloud cover = %f).",
                     weathHist[year].cloudcov_daily[doy]
                 );
                 // Will exit function prematurely due to error
@@ -2037,16 +1947,11 @@ void checkAllWeather(
                 // Otherwise, check if wind speed is less than 0 m/s
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Invalid daily wind speed value is less than zero."
-                    "(wind speed = %f) on day %d of year %d. ",
-                    weathHist[year].windspeed_daily[doy],
-                    doy + 1,
-                    year + weather->startYear
+                    "Daily wind speed value less than zero (wind speed = %f).",
+                    weathHist[year].windspeed_daily[doy]
                 );
                 // Will exit function prematurely due to error
 
@@ -2055,17 +1960,12 @@ void checkAllWeather(
                 // Otherwise, check if radiation if less than 0 W/m^2
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Invalid daily shortwave radiation value is less than "
-                    "zero."
-                    "(shortwave radation = %f) on day %d of year %d. ",
-                    weathHist[year].shortWaveRad[doy],
-                    doy + 1,
-                    year + weather->startYear
+                    "Daily shortwave radiation value less than zero. "
+                    "(shortwave radation = %f).",
+                    weathHist[year].shortWaveRad[doy]
                 );
                 // Will exit function prematurely due to error
 
@@ -2075,18 +1975,12 @@ void checkAllWeather(
                 // kPa
 
                 // Fail
-                LogErrorSuid(
+                LogError(
                     LogInfo,
                     LOGERROR,
-                    ncSuid,
-                    sDom,
-                    "Invalid daily actual vapor pressure value is less "
-                    "than "
-                    "zero."
-                    "(actual vapor pressure = %f) on day %d of year %d. ",
-                    weathHist[year].actualVaporPressure[doy],
-                    doy + 1,
-                    year + weather->startYear
+                    "Daily actual vapor pressure value less than zero. "
+                    "(actual vapor pressure = %f).",
+                    weathHist[year].actualVaporPressure[doy]
                 );
                 // Will exit function prematurely due to error
             }
@@ -2096,6 +1990,8 @@ void checkAllWeather(
             }
         }
     }
+
+    LogInfo->logDate[0] = '\0';
 }
 
 /**
@@ -2308,11 +2204,9 @@ void SW_WTH_new_day(
         LogError(
             LogInfo,
             LOGERROR,
-            "Missing weather data (day %u - %u) during simulation: "
+            "Missing weather: "
             "Tavg=%.2f, ppt=%.2f, wind=%.2f, rH=%.2f, vp=%.2f, rsds=%.2f / "
-            "cloud=%.2f\n",
-            year,
-            doy,
+            "cloud=%.2f",
             allHist[yearIndex].temp_avg[doy0],
             allHist[yearIndex].ppt[doy0],
             allHist[yearIndex].windspeed_daily[doy0],